Bugfix: Create patch parameters for every neuron individually
Now the patches are created with parameters that are defined for
every neuron individually.

Added function to create shared and partially overlapping patches.

Added docstrings and comments to every function.

Added citation of the paper in the beginning of the script.

Added return statements for all long-range connection functions to
plot connections and to facilitate debugging