Update: Added functions for direct stimulus reconstruction and fft
The idea of direct stimulus reconstruction is to use the information
about the color sensitivity to color the respective areas that
correspond to the receptive fields. The neurons with the lowest
firing rate are used first for dying the areas and the neurons with
most activity last. That ensures that the most present color is dominant
in the area.

Added also simple wrapper for a fast Fourier transform