Update: Added flags whether costly adj mat calc or weight conf is necessary
Computation of the adjacency matrix of sensory to sensory neurons is
cost intense but not always necessary. Similiarily, the configuration of
the synaptic weights in dependence on the number of the extablished
connections. The flags permit that these computations are left out.
Synaptic weights that are set via parameter cap_s are still set