Update: Incorporate inhibitory neurons into sensory layer and change Perlin discretisation
Inhibitory neurons are incorporated into the same sensory layer. Thus, it needs to handled
when synapses are established: synapses from inhibitory neurons have a negative weight.

Moreover, the Perlin discretisation takes the different distribution of values into account
and instead divides it by the number of values, such that every stimulus tuning class has
roughly the same number of values in the map