Update: Implemented function to have same input current for all neurons
To determine the magnitude of the input it is necessary to know
how the neurons react if all receive the same input. Naturally,
they show full synchrony, meaning that the transmission of signals
through recurrent connections does not have any effect due to the
refactory period.

Moreover, implemented update, such that image can be changed dynamically.

All classes accept kwargs for easy addition of new parameters