Bugfix: Connections established via nest connect
To avoid any connection to nodes that are no sensory neurons
(e.g. spike detector, mutlimeter etc) the connections are established
through the node tree and only with the sensory neurons. Although this
slows down the algorithm a little it makes the network behavior more
predictable.

Moreover, implemented functions to set the input current for the neurons