Update: Added function for mutual information of histograms
Renamed the original function to mutual_information_image since
it uses the image values (original and reconstructed) to compute
the mutual information. The new function uses the input stimulus and
the firing rates and compares the histograms of that. That is another
measurement for mutual information that does not require reconstruction
methods with non-linearities